app/routers/ai.py
import google.generativeai as genai
from fastapi import APIRouter, Depends, HTTPException, status
from app.schemas import GenerateImageRequest, AnalyzeImageRequest
from app.dependencies import get_current_user, get_supabase
from app.config import GOOGLE_API_KEY
from supabase import Client
import os

# Configure Gemini
if GOOGLE_API_KEY:
    genai.configure(api_key=GOOGLE_API_KEY)

router = APIRouter(prefix="/ai", tags=["AI"])

@router.post("/generate")
async def generate_image(
    request: GenerateImageRequest,
    current_user = Depends(get_current_user),
    supabase: Client = Depends(get_supabase)
):
    # This is a placeholder for actual Gemini Image Generation
    # Gemini Pro Vision is text/image-to-text. 
    # For Image Generation, we might need Imagen or a specific Gemini capability if available.
    # Assuming "gemini-3-pro-image-preview" is the model user mentioned.
    
    try:
        # 1. Fetch Business Profile context if needed
        profile_res = supabase.table("business_profiles").select("*").eq("id", current_user.id).single().execute()
        business_context = ""
        if profile_res.data:
            p = profile_res.data
            business_context = f"Brand: {p.get('business_name')}. Vibe: {p.get('vibes')}. Theme: {p.get('theme')}."

        # 2. Fetch Dataset context if provided
        dataset_context = ""
        if request.dataset_id:
             dataset_res = supabase.table("datasets").select("*").eq("id", request.dataset_id).single().execute()
             if dataset_res.data:
                 dataset_context = f"Style Guidelines: {dataset_res.data.get('master_prompt')}"

        full_prompt = f"{business_context} {dataset_context} {request.prompt}"
        
        # Placeholder for actual generation call
        # model = genai.GenerativeModel('gemini-3-pro-image-preview') # Hypothetical model name
        # response = model.generate_content(full_prompt)
        
        # Mock response for now as we don't have the exact model capability in this env
        # In a real scenario, you'd get the image, upload to Supabase 'generated-images', and return URL.
        
        return {"message": "Image generation initiated", "prompt_used": full_prompt}
        
    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))

@router.post("/analyze")
async def analyze_style(
    request: AnalyzeImageRequest,
    current_user = Depends(get_current_user),
):
    try:
        model = genai.GenerativeModel('gemini-pro-vision')
        
        # In reality, we'd need to download the images from URLs and pass them to the model
        # prompt = "Analyze these images and extract a master style prompt including lighting, colors, and vibe."
        
        return {"message": "Analysis endpoint ready (requires image download logic)"}
    except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
