requirements.txt
google-generativeai
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
email-validator
python-multipart
